+++ packages/shared/src/UI/components/PersonaAction/index.ts
@@ -1,3 +1,3 @@
-export * from './ImageIcon'
+export * from './PersonaImageIcon'
 export * from './PersonaAction'
 export * from './PlatformAvatar'
