+++ packages/shared/src/utils/index.ts
@@ -68,5 +68,6 @@ export function createNonFungibleToken<ChainId, SchemaType>(
 }
 
 export const formatPublicKey = (publicKey?: string) => {
-    return `${publicKey?.slice(0, 6)}...${publicKey?.slice(-6)}`
+    if (!publicKey) return ''
+    return `${publicKey.slice(0, 6)}...${publicKey.slice(-6)}`
 }
